Add min_node_count argument and test
This new argument says "You must see at least this many nodes exit before
you consider the job done".

This is necessary because the current logic of checking the number of nodes
acked against the number of jobs exited doesn't account for a job acking
and exiting before another job has even acked, which is easy to trigger.